cleaned caps: do not strip type specifier
For one the field colorimetry requires a type of (string) on the command.
The other way around is to parse the fields and special case each fields
to restore the type specifier for the fields taht requires it, here colorimetry.

Though this seems overkill without a rationale for stripping the type
specifiers.