cleaned caps: do not strip type specifier, escape them
For one the field colorimetry requires a type of (string) on the command.
The other way around is to parse the fields and special case each fields
to restore the type specifier for the fields taht requires it, here colorimetry.

For now it looks overkill to build a parser, so escape the parens around type
specifiers.