Make extension work with Wayland
Currently the extension prevents the Gnome Wayland session
from starting. This seems to be caused by GStreamer initializing
too early.
This commit delays the calls to GStreamer in the extension and
the preferences until the user interacts with the widgets. At that
time GStreamer is ready to be called and the extension works as
expected.