Use the standard ByteArray.toString conversion
This addresses the following warning:

Some code called array.toString() on a Uint8Array instance. Previously
this would have interpreted the bytes of the array as a string, but that
is nonstandard. In the future this will return the bytes as
comma-separated digits. For the time being, the old behavior has been
preserved, but please fix your code anyway to explicitly call
ByteArray.toString(array).

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>